requirements.txt
# Python Dependencies for Smart DentalOps Backend

# FastAPI - Modern web framework for building APIs
fastapi>=0.104.0

# Uvicorn - ASGI server to run FastAPI
uvicorn[standard]>=0.24.0

# Pandas - Data manipulation and CSV handling
pandas>=2.0.0

# NumPy - Numerical computing
numpy>=1.24.0

# Scikit-learn - Machine learning utilities (includes MLPClassifier for neural networks)
scikit-learn>=1.3.0

# Joblib - For saving/loading models
joblib>=1.3.0

# Pydantic - Data validation (comes with FastAPI but listed for clarity)
pydantic>=2.0.0

# Python-multipart - For handling form data (if needed later)
python-multipart>=0.0.6

# PHASE 2: Sentiment Analysis
# TextBlob - Simple sentiment analysis library
textblob>=0.17.1

# PHASE 4: Authentication
# JWT token handling
python-jose[cryptography]>=3.3.0

# Email validation
email-validator>=1.1.5

